cli: factorize the train command on every dataset
