+++ kpconv_torch/train.py
@@ -6,155 +6,45 @@ import time
 import numpy as np
 from torch.utils.data import DataLoader
 
-from kpconv_torch.datasets.S3DIS import S3DISCollate, S3DISDataset, S3DISSampler
-from kpconv_torch.models.architectures import KPFCNN
-from kpconv_torch.utils.config import Config
+from kpconv_torch.datasets.ModelNet40 import (
+    ModelNet40Collate,
+    ModelNet40Config,
+    ModelNet40Dataset,
+    ModelNet40Sampler,
+)
+from kpconv_torch.datasets.NPM3D import (
+    NPM3DCollate,
+    NPM3DConfig,
+    NPM3DDataset,
+    NPM3DSampler,
+)
+from kpconv_torch.datasets.S3DIS import (
+    S3DISCollate,
+    S3DISConfig,
+    S3DISDataset,
+    S3DISSampler,
+)
+from kpconv_torch.datasets.SemanticKitti import (
+    SemanticKittiCollate,
+    SemanticKittiConfig,
+    SemanticKittiDataset,
+    SemanticKittiSampler,
+)
+from kpconv_torch.datasets.Toronto3D import (
+    Toronto3DCollate,
+    Toronto3DConfig,
+    Toronto3DDataset,
+    Toronto3DSampler,
+)
+from kpconv_torch.models.architectures import KPCNN, KPFCNN
 from kpconv_torch.utils.trainer import ModelTrainer
-
-
-class S3DISConfig(Config):
-    """
-    Override the parameters you want to modify for this dataset
-    """
-
-    ####################
-    # Dataset parameters
-    ####################
-    # Dataset name
-    dataset = "S3DIS"
-
-    # Number of classes in the dataset (This value is overwritten by dataset class when Initializating dataset).
-    num_classes = None
-
-    # Type of task performed on this dataset (also overwritten)
-    dataset_task = ""
-
-    # Number of CPU threads for the input pipeline
-    input_threads = 10
-
-    #########################
-    # Architecture definition
-    #########################
-    # # Define layers
-    architecture = [
-        "simple",
-        "resnetb",
-        "resnetb_strided",
-        "resnetb",
-        "resnetb",
-        "resnetb_strided",
-        "resnetb",
-        "resnetb",
-        "resnetb_strided",
-        "resnetb_deformable",
-        "resnetb_deformable",
-        "resnetb_deformable_strided",
-        "resnetb_deformable",
-        "resnetb_deformable",
-        "nearest_upsample",
-        "unary",
-        "nearest_upsample",
-        "unary",
-        "nearest_upsample",
-        "unary",
-        "nearest_upsample",
-        "unary",
-    ]
-
-    ###################
-    # KPConv parameters
-    ###################
-    # Number of kernel points
-    num_kernel_points = 15
-
-    # Radius of the input sphere (decrease value to reduce memory cost)
-    in_radius = 1.2
-
-    # Size of the first subsampling grid in meter (increase value to reduce memory cost)
-    first_subsampling_dl = 0.03
-
-    # Radius of convolution in "number grid cell". (2.5 is the standard value)
-    conv_radius = 2.5
-
-    # Radius of deformable convolution in "number grid cell". Larger so that deformed kernel can spread out
-    deform_radius = 5.0
-
-    # Radius of the area of influence of each kernel point in "number grid cell". (1.0 is the standard value)
-    KP_extent = 1.2
-
-    # Behavior of convolutions in ('constant', 'linear', 'gaussian')
-    KP_influence = "linear"
-
-    # Aggregation function of KPConv in ('closest', 'sum')
-    aggregation_mode = "sum"
-
-    # Choice of input features
-    first_features_dim = 128
-    in_features_dim = 5
-
-    # Can the network learn modulations
-    modulated = False
-
-    # Batch normalization parameters
-    use_batch_norm = True
-    batch_norm_momentum = 0.02
-
-    # Deformable offset loss
-    # 'point2point' fitting geometry by penalizing distance from deform point to input points
-    # 'point2plane' fitting geometry by penalizing distance from deform point to input point triplet (not implemented)
-    deform_fitting_mode = "point2point"
-    deform_fitting_power = 1.0  # Multiplier for the fitting/repulsive loss
-    deform_lr_factor = 0.1  # Multiplier for learning rate applied to the deformations
-    repulse_extent = 1.2  # Distance of repulsion for deformed kernel points
-
-    #####################
-    # Training parameters
-    #####################
-    # Maximal number of epochs
-    max_epoch = 500
-
-    # Learning rate management
-    learning_rate = 1e-2
-    momentum = 0.98
-    lr_decays = {i: 0.1 ** (1 / 150) for i in range(1, max_epoch)}
-    grad_clip_norm = 100.0
-
-    # Number of batch (decrease to reduce memory cost, but it should remain > 3 for stability)
-    batch_num = 6
-
-    # Number of steps per epochs
-    epoch_steps = 500
-
-    # Number of validation examples per epoch
-    validation_size = 50
-
-    # Number of epoch between each checkpoint
-    checkpoint_gap = 50
-
-    # Augmentations
-    augment_scale_anisotropic = True
-    augment_symmetries = [True, False, False]
-    augment_rotation = "vertical"
-    augment_scale_min = 0.9
-    augment_scale_max = 1.1
-    augment_noise = 0.001
-    augment_color = 0.8
-
-    # The way we balance segmentation loss
-    #   > 'none': Each point in the whole batch has the same contribution.
-    #   > 'class': Each class has the same contribution (points are weighted according to class balance)
-    #   > 'batch': Each cloud in the batch has the same contribution (points are weighted according cloud sizes)
-    segloss_balance = "none"
-
-    # Do we nee to save convergence
-    saving = True
-    saving_path = None
 
 
 def main(args):
     ############################
     # Initialize the environment
     ############################
+    start = time.time()
     # Set which gpu is going to be used
     GPU_ID = "0"
 
@@ -166,7 +56,7 @@ def main(args):
     if args.chosen_log:
 
         # Find all snapshot in the chosen training folder
-        chkp_path = os.path.join("results", args.chosen_log, "checkpoints")
+        chkp_path = os.path.join(args.chosen_log, "checkpoints")
         chkps = [f for f in os.listdir(chkp_path) if f[:4] == "chkp"]
 
         # Find which snapshot to restore
@@ -174,9 +64,7 @@ def main(args):
             chosen_chkp = "current_chkp.tar"
         else:
             chosen_chkp = np.sort(chkps)[chkp_idx]
-        chosen_chkp = os.path.join(
-            "results", args.chosen_log, "checkpoints", chosen_chkp
-        )
+        chosen_chkp = os.path.join(args.chosen_log, "checkpoints", chosen_chkp)
 
     else:
         chosen_chkp = None
@@ -189,33 +77,86 @@ def main(args):
     print("****************")
 
     # Initialize configuration class
-    config = S3DISConfig()
+    if args.dataset == "ModelNet40":
+        config = ModelNet40Config()
+    if args.dataset == "NPM3D":
+        config = NPM3DConfig()
+    if args.dataset == "S3DIS":
+        config = S3DISConfig()
+    if args.dataset == "SemanticKitti":
+        config = SemanticKittiConfig()
+    elif args.dataset == "Toronto3D":
+        config = Toronto3DConfig()
+
     if args.chosen_log:
-        config.load(os.path.join("results", args.chosen_log))
+        config.load(args.chosen_log)
+        if config.dataset != args.dataset:
+            raise ValueError(
+                f"Config dataset ({config.dataset}) "
+                f"does not match provided dataset ({args.dataset})."
+            )
         config.saving_path = None
 
     # Get path from argument if given
     if len(sys.argv) > 1:
         config.saving_path = sys.argv[1]
 
-    # Initialize datasets
-    training_dataset = S3DISDataset(
-        args.datapath, config, set="training", use_potentials=True
-    )
-    test_dataset = S3DISDataset(
-        args.datapath, config, set="validation", use_potentials=True
-    )
-
-    # Initialize samplers
-    training_sampler = S3DISSampler(training_dataset)
-    test_sampler = S3DISSampler(test_dataset)
+    # Initialize datasets and samplers
+    if config.dataset == "ModelNet40":
+        training_dataset = ModelNet40Dataset(args.datapath, config, train=True)
+        test_dataset = ModelNet40Dataset(args.datapath, config, train=False)
+        training_sampler = ModelNet40Sampler(training_dataset, balance_labels=True)
+        test_sampler = ModelNet40Sampler(test_dataset, balance_labels=True)
+        collate_fn = ModelNet40Collate
+    elif config.dataset == "NPM3D":
+        training_dataset = NPM3DDataset(
+            args.datapath, config, set="training", use_potentials=True
+        )
+        test_dataset = NPM3DDataset(
+            args.datapath, config, set="validation", use_potentials=True
+        )
+        training_sampler = NPM3DSampler(training_dataset)
+        test_sampler = NPM3DSampler(test_dataset)
+        collate_fn = NPM3DCollate
+    elif config.dataset == "S3DIS":
+        training_dataset = S3DISDataset(
+            args.datapath, config, set="training", use_potentials=True
+        )
+        test_dataset = S3DISDataset(
+            args.datapath, config, set="validation", use_potentials=True
+        )
+        training_sampler = S3DISSampler(training_dataset)
+        test_sampler = S3DISSampler(test_dataset)
+        collate_fn = S3DISCollate
+    elif config.dataset == "SemanticKitti":
+        training_dataset = SemanticKittiDataset(
+            args.datapath, config, set="training", balance_classes=True
+        )
+        test_dataset = SemanticKittiDataset(
+            args.datapath, config, set="validation", balance_classes=False
+        )
+        training_sampler = SemanticKittiSampler(training_dataset)
+        test_sampler = SemanticKittiSampler(test_dataset)
+        collate_fn = SemanticKittiCollate
+    elif config.dataset == "Toronto3D":
+        training_dataset = Toronto3DDataset(
+            args.datapath, config, set="training", use_potentials=True
+        )
+        test_dataset = Toronto3DDataset(
+            args.datapath, config, set="validation", use_potentials=True
+        )
+        training_sampler = Toronto3DSampler(training_dataset)
+        test_sampler = Toronto3DSampler(test_dataset)
+        collate_fn = Toronto3DCollate
+    else:
+        raise ValueError("Unsupported dataset : " + config.dataset)
 
     # Initialize the dataloader
     training_loader = DataLoader(
         training_dataset,
         batch_size=1,
         sampler=training_sampler,
-        collate_fn=S3DISCollate,
+        collate_fn=collate_fn,
         num_workers=config.input_threads,
         pin_memory=True,
     )
@@ -223,10 +164,15 @@ def main(args):
         test_dataset,
         batch_size=1,
         sampler=test_sampler,
-        collate_fn=S3DISCollate,
+        collate_fn=collate_fn,
         num_workers=config.input_threads,
         pin_memory=True,
     )
+
+    if config.dataset == "SemanticKitti":
+        # Calibrate max_in_point value
+        training_sampler.calib_max_in(config, training_loader, verbose=True)
+        test_sampler.calib_max_in(config, test_loader, verbose=True)
 
     # Calibrate samplers
     training_sampler.calibration(training_loader, verbose=True)
@@ -242,7 +188,12 @@ def main(args):
 
     # Define network model
     t1 = time.time()
-    net = KPFCNN(config, training_dataset.label_values, training_dataset.ignored_labels)
+    if config.dataset == "ModelNet40":
+        net = KPCNN(config)
+    else:
+        net = KPFCNN(
+            config, training_dataset.label_values, training_dataset.ignored_labels
+        )
 
     debug = False
     if debug:
@@ -271,3 +222,6 @@ def main(args):
 
     print("Forcing exit now")
     os.kill(os.getpid(), signal.SIGINT)
+
+    end = time.time()
+    print(time.strftime("%H:%M:%S", time.gmtime(end - start)))
